refactor: FileBrowserPaneView のドラッグ&ドロップ処理を専用ハンドラに分離 (#157)
内部項目の移動/コピー・外部ファイル/フォルダの受け入れ・Explorer 等へのドラッグアウト・
ブレッドクラムへのドロップと、ヒットテスト（IsInternalDrag / TryGetDropTargetFolder /
TryGetBreadcrumbTarget）・ドラッグ状態（_dragItems / _wasInternalDrop / InternalDragKey）を
Panes/FileBrowser/FileBrowserDragDropHandler.cs（internal sealed）へ移動した。

D&D はビジュアルツリー操作・DataPackage 操作を伴う純粋な View 責務のため、ViewModel では
なく View 層ヘルパーとして分離（MVVM ガードレール準拠）。ハンドラはコンストラクタで
ヒットテスト用ルート要素（RootGrid）・ViewModel アクセサ・FileBrowserDialogs を受け取る。
XAML のイベント結線は View に薄い委譲ハンドラとして残し、処理本体をハンドラへ移譲した
（表示・操作挙動は不変）。

FindAncestor<T> は View の他箇所でも使用するため View に残し、D&D 専用の IsDescendantOf は
ハンドラへ移動。FileBrowserPaneView.xaml.cs を 1,515 行から 1,191 行へ削減（−324 行）。
build / test（608 件合格）/ format いずれも Release x64 で確認。

Closes #157
